prompt-engineering-kr/requirements.txt
# AI API 클라이언트
openai>=1.0.0
google-generativeai>=0.3.0

# 환경 설정 및 HTTP 요청
python-dotenv>=1.0.0
requests>=2.31.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 시각화 (선택적)
matplotlib>=3.7.0
seaborn>=0.12.0

# 유틸리티
tqdm>=4.66.0
colorama>=0.4.6

# 개발 및 테스트 도구 (선택적)
pytest>=7.4.0
black>=23.7.0
pylint>=2.17.0